refactor: Extract shouldReturnResultWithSuccessWhenQueryIsIssn
