+++ src/test/java/no/sikt/nva/pubchannels/handler/search/BaseSearchSerialPublicationByQueryHandlerTest.java
@@ -29,6 +29,7 @@ import no.unit.nva.commons.pagination.PaginatedSearchResult;
 import no.unit.nva.stubs.FakeContext;
 import nva.commons.apigateway.GatewayResponse;
 import nva.commons.apigateway.exceptions.UnprocessableContentException;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
@@ -47,33 +48,58 @@ public abstract class BaseSearchSerialPublicationByQueryHandlerTest extends Sear
         var year = randomYear();
         var issn = randomIssn();
         var testChannel = new TestChannel(year, UUID.randomUUID().toString(), type).withPrintIssn(issn);
-        mockChannelRegistryResponse(year, issn, List.of(testChannel.asChannelRegistrySerialPublicationBody()));
         var input = constructRequest(Map.of("year", year, "query", issn), mediaType);
+        var expectedSearchResult = mockChannelFoundAndReturnExpectedResponse(year, issn, testChannel);
+        var expectedMediaType =
+            mediaType.equals(MediaType.ANY_TYPE) ? MediaType.JSON_UTF_8.toString() : mediaType.toString();
 
         this.handlerUnderTest.handleRequest(input, output, context);
         var response = GatewayResponse.fromOutputStream(output, PaginatedSearchResult.class);
         var pagesSearchResult = objectMapper.readValue(response.getBody(), TYPE_REF);
         var contentType = response.getHeaders().get(CONTENT_TYPE);
 
-        var expectedMediaType =
-            mediaType.equals(MediaType.ANY_TYPE) ? MediaType.JSON_UTF_8.toString() : mediaType.toString();
-        var expectedSearchResult = getExpectedSearchResult(year, issn, testChannel);
         assertThat(pagesSearchResult.getHits(), containsInAnyOrder(expectedSearchResult.getHits().toArray()));
         assertThat(contentType, is(equalTo(expectedMediaType)));
         assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_OK)));
     }
 
-    private PaginatedSearchResult<SerialPublicationDto> getExpectedSearchResult(String year,
-                                                                                String printIssn,
-                                                                                TestChannel testChannel)
+    @Test
+    void shouldReturnResultWithSuccessWhenQueryIsIssn() throws IOException, UnprocessableContentException {
+        var year = randomYear();
+        var issn = randomIssn();
+        var expectedSearchResult = getExpectedPaginatedSearchResultIssnSearch(year, issn);
+
+        var input = constructRequest(Map.of("year", year, "query", issn), MediaType.ANY_TYPE);
+
+        this.handlerUnderTest.handleRequest(input, output, context);
+
+        var response = GatewayResponse.fromOutputStream(output, PaginatedSearchResult.class);
+        var pagesSearchResult = objectMapper.readValue(response.getBody(), TYPE_REF);
+
+        assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_OK)));
+        assertThat(pagesSearchResult.getHits(), containsInAnyOrder(expectedSearchResult.getHits().toArray()));
+    }
+
+    private PaginatedSearchResult<SerialPublicationDto> getExpectedPaginatedSearchResultIssnSearch(
+        String year, String printIssn)
+        throws UnprocessableContentException {
+        var pid = UUID.randomUUID().toString();
+        var testChannel = new TestChannel(year, pid, type).withPrintIssn(printIssn);
+        return mockChannelFoundAndReturnExpectedResponse(year, printIssn, testChannel);
+    }
+
+    private PaginatedSearchResult<SerialPublicationDto> mockChannelFoundAndReturnExpectedResponse(String year,
+                                                                                                  String issn,
+                                                                                                  TestChannel testChannel)
         throws UnprocessableContentException {
         var expectedParams = new HashMap<String, String>();
-        expectedParams.put("query", printIssn);
+        expectedParams.put("query", issn);
         if (nonNull(year)) {
             expectedParams.put("year", year);
         }
 
         var expectedHits = List.of(testChannel.asSerialPublicationDto(selfBaseUri, year));
+        mockChannelRegistryResponse(year, issn, List.of(testChannel.asChannelRegistrySerialPublicationBody()));
 
         return PaginatedSearchResult.create(constructPublicationChannelUri(testChannel.type(), expectedParams),
                                             DEFAULT_OFFSET_INT,
