+++ src/test/java/no/sikt/nva/pubchannels/handler/search/series/SearchSeriesByQueryHandlerTest.java
@@ -68,23 +68,6 @@ class SearchSeriesByQueryHandlerTest extends BaseSearchSerialPublicationByQueryH
     }
 
     @Test
-    void shouldReturnResultWithSuccessWhenQueryIsIssn() throws IOException, UnprocessableContentException {
-        var year = randomYear();
-        var issn = randomIssn();
-        var expectedSearchResult = getExpectedPaginatedSearchResultIssnSearch(year, issn);
-
-        var input = constructRequest(Map.of("year", year, "query", issn), MediaType.ANY_TYPE);
-
-        this.handlerUnderTest.handleRequest(input, output, context);
-
-        var response = GatewayResponse.fromOutputStream(output, PaginatedSearchResult.class);
-        var pagesSearchResult = objectMapper.readValue(response.getBody(), TYPE_REF);
-
-        assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_OK)));
-        assertThat(pagesSearchResult.getHits(), containsInAnyOrder(expectedSearchResult.getHits().toArray()));
-    }
-
-    @Test
     void shouldReturnResultWithRequestedYearIfThirdPartyDoesNotProvideYear()
         throws IOException, UnprocessableContentException {
         var year = randomYear();
@@ -271,14 +254,6 @@ class SearchSeriesByQueryHandlerTest extends BaseSearchSerialPublicationByQueryH
 
     private static SerialPublicationDto toResult(ThirdPartySerialPublication series, String requestedYear) {
         return SerialPublicationDto.create(constructPublicationChannelUri(SERIES_PATH, null), series, requestedYear);
-    }
-
-    private PaginatedSearchResult<SerialPublicationDto> getExpectedPaginatedSearchResultIssnSearch(
-        String year, String printIssn)
-        throws UnprocessableContentException {
-        var pid = UUID.randomUUID().toString();
-        var testChannel = new TestChannel(year, pid, "Series").withPrintIssn(printIssn);
-        return createSearchResult(testChannel, String.valueOf(year), printIssn);
     }
 
     private PaginatedSearchResult<SerialPublicationDto> getExpectedPaginatedSearchResultIssnSearchThirdPartyDoesNotProvideYear(
